[#3] Add a CI config and corresponding files
Problem: we want CI to check that various properties about
this repo (e. g. that code compiles).

Solution: add a CI config for buildkite which is the service we
use most often. Add some files necessary to make it work (configs,
scripts).
Also add a simple .bats test for the `baseDAO` executable.