[#29] Make decision lambda more practical and safe
Problem:

The current decision lambda has too few sensible use cases.

It can modify the contract storage, read the proposal and perform
operations.
Regarding operations it is clear, but there is nothing in the storage
that we could want to update.

Another problem is that `decisionLambda` can change anything in the
storage and thus break the contract's invariants.

Solution:

1. Add second `contractMetadata` type parameter to config and storage,
and store value of this type in the storage.

This way user can gather interesting global info based on accepted
proposals.

2. Change decision lambda so that it can read the proposal, read
storage, modify only its `contractMetadata` part and perform operations.

This version is safer as we do not allow changing to the storage.

We decided that it does not make much sense to modify proposals
because:

(1) updating direct fields of `Proposal` is apparently not proper as
this can break inner contract invariants.

(2) if we modify some of the existing fields of `proposalMetadata`, that
does not look proper because this way we (contract author) overwrite
user's data.

(3) if we add a dedicated field to `proposalMetadata` to be filled after
proposal being accepted, that would mean that this field remains unused
during proposal being ongoing, and it is not clear how to initialize
this field on proposal creation. In other words, this is not idiomatic.
We suggest using `contractMetadata` space for such purposes.

(4) finally, we use hash of the proposal as key for proposals map, and
modifying the proposal would bring quite a chaos into the contract.