[#29] Remove `emptyStorage`
Problem: it seems unnecessary. It is not usable in production (due to
`admin=genesisAddress`), and in tests we can't use it for the same
reason because Nettest does not work with `genesisAddress`.

It also served as an example of initializing storage, but soon we are
going to add DAO template contract which fits for this purpose better.

Solution: remove `emptyStorage`.