[#38] Extract `unfreezeProposerToken` to stack
Problem: this function is used in three various places and is quite big,
consuming about `1300` bytes in origination operation. We would like to
spare some bits.

Solution: push this function once on stack in the beginning and call
from there.

We have already did this with `creditTo` and `debitFrom`, however, here
situation is much more complex - `unfreezeProposerToken` has to call
`creditTo` and `debitFrom` inside. So we need one lambdas on stack to
call other ones.

This commit assumes the following workflow: `unfreezeProposerToken`'s
implementation accepts `creditTo` and `debitFrom` lambdas explicitly.
Once pushed on stack at the beginning, we use `apply` (partial
application for lambdas) to feed those two functions to our complex
function, and then we call it like if it had no dependencies.