[#31] Generalize origination return value in tests
Problem: I need to add one more return value in `originateBaseDao*`
methods, however this means updating all the existing tests. This should
not work such way.

Solution:

One straightforward solution would be returning a datatype and picking
the necessary fields like `TestSetup{ tsDao = dao, ... }`, and we apply
it here partially.

The problem with it is that we try to avoid boilerplate, and it would be
nice to leave the current shortness of tests. To achieve this, we
introduce a typeclass fo injections (`>->`), and let the user to pick
various parts of `TestSetup` using `dao :/ part2 :/ part3` syntax.

Here we have to rely on types to understand which piece user wants to
pick. While contract address uses a quite unique type (`TAddress`),
administrator and owners can be confused. Unfortunatelly, it is not
possible to pick a thing by name in patterns, nor pick one of many
values (e.g. only one owner like `owner1 : _`) without requiring
`MonadFail`. So with this approach we are quite constrained in how
to pick the necessary parts with low boilerplate.