[#69] Allow `flush` to accepts an amout of proposals, add `dropProposals` entrypoint
Problem: When calling `flush`, `decisionLambda` or `rejectedValue`
lambda may fail due to unexpected error. We should have a way to drop
some proposals and a way to not flush proposals all at one.

Solution:
- Refactor `sProposalKeyListSortByDate` to be a map with
`(Timestamp, ProposalKey proposalMetadata)` as the key, and `()` as
the value (Since value here is not important).
- Add `dropProposals` entrypoint
- Update `flush` entrypoint to accept a (Maybe Natural)
- Add tests for these new features