[#38] Revert `unfreezeProposerToken` extraction
Problem: it turned out that this change increased the origination
operation size. It happens because printing the signature of extracted
lambda costs more than 700 bytes (it contains signatures of `creditTo`
and `debitFrom`, and they all mention storage type which is big).

Solution: revert those changes, call `unfreezeProposerToken` inlined.

I suppose that in overall extracting lambdas depending on other lambdas
is not viable - unless for very big pieces of code, so removing the made
changes entirely.