[#57] Add first tests in LIGO version of baseDAO
Problem: we want to initialize tests package for LIGO implementation.

Solution: add a test package in `ligo` folder with cleveland tests.

For now I add only one test - that contract can be imported.

We have to redefine some types for this to work, since LIGO tends to
reorder fields and constructors, and probably we don't want to do this
in our Parameter, Storage and Config. So I mirror them in test package
and provide convertions.