[#100] Adding Haskell tests for management entrypoints
Problem: Tests that check the behavior of management entrypoints is
missing. We should add them.

Solution: Add the tests. The tests for the management entrypoints has
already been implemented for the Lorentz version of the contract. Some
attempt was made to reuse them without duplication, but didn't quite
workout. So these tests have been copied here, with an altered version
of 'withOriginated' function that originates the LIGO contract.

A new test for 'CallCustom' entrypoint was added to check that it can be
called as expected.