[#91] Refactor ligo and haskell parameters
Problem: We would like to check the migration status and XTZ transfer
for many entrypoints, but some entrypoints needs to be excluded as wel.
Because of flat structure of the parameter, we have to either duplicate
these tests for the entrypoints that require them. We would like to
avoid it.

Solution: Make a nested structure so that we can do such checks on a
one branch of a conditional, thus eliminating the need to duplicate the
function calls.