[#106] Update LIGO `failwith` to be the same as Lorentz version
Problem: Current LIGO only fails with type `string` which is different
from the Lorentz version. Because of this, we require a `isLorentz`
check in the tests to be able to share the tests properly. We can update
the `failwith` in LIGO to be the same as lorentz which would make the
check no longer necessary.

Solution: Update LIGO `failwith` to be the same as Lorentz version,
remove `isLorentz` check in share tests.