[#107] Remove token metadata from LIGO version
Problem: it is unfortune that token metadata now appears in LIGO's
contract storage. In Lorentz we didn't care because Config never
appeared in Michelson, but in LIGO version it does, and this metadata
increases the contract size substantially.

Solution: remove token metadata from there, since it is already provided
via an off-chain view.