[#123] Reduce lambdas origination cost by refactoring arguments
Problem: the contract is too big to fit into the limits for
origination, partly because the type signature for the lambdas in it
(the ones generated by LIGO, as re-used functions) are too big.

Solution: reduce the const of the overall contract by refactoring
the auto-generated lambdas to have smaller types as arguments.