[#116] Add RegistryDao decision lambda test
Problem: The tests for decision lambda are missing, because it is hard
to check for storage changes, since it contains bigmaps.

Solution: Implement a test for decision lambda behavior where we update
s_max value with a proposal, and test it by creating another proposal
which should only work if the `s_max` value was updated.