[#131] Remove `FAILWITH` `Pair "string" Unit` from LIGO
Problem: In many places we fail with `Pair "string" Unit`.
In case of the Lorentz version, it's standard behavior which
is not really justified and should be changed in Lorentz.

Solution: Use a `failAsText` helper in Lorentz to no longer fail with
`Pair "string" Unit`, refactor LIGO code to use standard `failwith`
where it can, make adjustment to all the tests.