[#123] Modify README and Makefile rule to use be entrypoint-specific
Problem: We currently create the entrypoints with a Makefile rule
to be loaded into the contract as a parameter to the whole contract,
but we could instead make them a parameter for the 'startup'
entrypoint, which would be easier to use.

Solution: modify Makefile and test to create/use the entrypoints as
parameters to 'startup' and update the README documentation.