[#137] Update LIGO Contract to add the new proposal type
Problem: We have to add a new proposal type that can add or remove items
from a list of addresses in storage.

Solution: Add a new field of type 'address set' to registryDAO storage
"proposal_receivers". Add a new proposal type that is identified by
"update_receivers" field in the proposal metadata. The packed bytes from
this field is unpacked into the 'update_receivers' type. The code
branches on this type to add or remove items from the Set of addresses.