[#137] Update Lorentz contract to add the new proposal type
Problem: We have to add the new proposal type to the Lorentz contract.

Solution: Add the new field 'ceProposalReceivers' to the
'RegistryDAOContract' type. Add a new constructor to
'RegistryDaoProposalMetadata' type, that wraps the type
'UpdateReceivers'. Add code to the main contract to handle this type of
proposal. Update tests to account for the change in length of proposal
metadata. Also update RegistryDAO sepecification.