[#137] Add sub-command to print metadata of registryDAO
Problem: Earlier the metadata was shared between all the contract, but
now registryDAO contract should have a slightly different metadata with
the TZIP-16 view to fetch the proposal receiver contract addresses. This
means that we should add a command for the user to print the Metadata of
registryDAO contract.

Solution: Add the CLI options to add such a command.