[#169] remove Ligo ShareTest from its Haskell library
Problem: there are ShareTest modules in the Haskell support library
for the Ligo version, however these are no longer shared and should
be in the tests only.

Solution: very much as it was done for the Lorentz version here the
ShareTest modules' code was merged inside the rest of the haskell test
modules.