[#169] deprecate Lorentz version
Problem: the Lorentz version needs to be deprecated in favor of the
Ligo version, but it is still shown and documented as the main one.

Solution: move the Lorentz version of the contract to its own
subdirectory and document its deprecated status.
Additionally, this updates the pipeline to improve its speed and
(due to a recent infrastructure update) correctly use multiple
workers.