[#175] make the quorum a fraction of Total Supply
Problem: currently the quorum needs to be adjusted through voting once
the total supply increases.
However we could instead have the quorum as a fraction of it, allowing
it to scale accordingly.

Solution: Modify the quorum threshold, as well as its settings, to be
a fraction of the total supply of frozen tokens.