[#170] Implement BYO Extension
Problem: We have to implement BYO extension, where an external FA2
contract is used to lock the tokens for users when they freeze tokens in
this contract to propose or vote.

Solution: Implement the changes in LIGO source. As part of this, we also
stopped tracking unfrozen tokens in this contract, and have removed
references to unfrozen token from LIGO contract. We also added a
`governance_token` type to wrap the address of the external contract and
the Token Id on that contract that we should operate on during
freeze/unfreeze operations. Makefile has been changed to accept the
attributes of this governance token during contract origination and
storage creation.