[#188] Store `contract_extra` as a `big_map`
Problem: The `extra` field of the storage is a `contract_extra`,
which is currently defined as a `(string, bytes) map`.
This however poses a limitation, as `map` is strict and so the cost
in case of many key is too high.

Solution: Make `contract_extra` as `big_map` instead of map, update
various unpack related functions.