[#171] Update storage to include voting_period_params
Problem: We need to sync the contract with Tezos periods and cycles. For
this we only have the `level` instruction to use.

Solution: To compute the current tezos period and cycle, we use the
current level with a known period with known starting level and cycle.
These known values need to be stored in the storage. So we define a
record to hold these values and include it in the storage. We also
define ligo functions to compute current cycle and period from the
current level and the stored constants.