[#171] Update freeze/unfreeze processes to implement VotingPowerDAO
Problem: To implement VotingPowerDAO logic, we update freeze, unfreeze
entrypoint procedures to consider VOTING_POWER of the accounts to limit
the number of tokens that can be freezed. Also we have to consider
TOTAL_VOTING_POWER as the total_supply while calculating the quorum
value.

Solution: Update proposal.mligo to implement the required logic. As part
of this we have changed the entrypoint of `freeze` to include the
key_hash of the sender. This was done because there appear to be no way
to derive key_hash from address. The unfreeze entrypoint now just checks
that there are enough unstaked frozen_tokens in the current period and
just burns them. The proposal quorum calculation was changed to consider
TOTAL_VOTING_POWER as denominator while calculating proposal quorum.
In addition to this the type of the decision lambda was changed to
include the voting_period_params in the return, as a means to update
the voting_period_params stored in storage, when they require changing.
The registryDAO and treasuryDAO was updated to handle the new proposal
type.