[#213] Updating tests
Problem: We have removed 'set_quorum_threshold' entrypoint from contract
interface. Now we have to update tests that use this entrypoint.

Solution: Remove existing tests that use the entrypoint. As part of this
change, the 'quorum_threshold' setting have been moved to 'config' part
of the storage in the main contract.