[#210] Remove the special `admin` permissions on `transfer`
Problem: The `transfer` entrypoint has a special permission that
allows the `admin` (and only him) to transfer frozen tokens on the ledger.
This was intended as a shorthand for `mint` and `burn`, which granted him
the same ability. However, since these have been removed (see #208 and #209) we should
remove this permission as well.

Solution: Remove the special permission which allow `admin` to transfer
frozen token.