[#208] Remove the `mint` entrypoint
Problem: `mint` is an admin-only entrypoint that allow
generating tokens. This was intended to
be used mainly to set up the contract and then to not be used again.
However, it is possible to setup the ledger with correct initial value
at origination and we no longer need this entrypoint.

Solution: Remove `mint` entrypoint and remove all related
tests.