[#203] Add test to check that operators are token-specific
Problem: the current implementation does not save the token_id for
operators, so an operator for a token holder is able to operate
every custom token that the DAO might have.
This however is not in line with the FA2 specification.

Solution: add a test to check the expected property and verify
that it fails.