[#238] Remove morley-large-originator
Problem: With the activation of `florence` the limit for the origination contract
size is much higher and it should be possible to not need `morley-large-originator` to do so.
For the context:
- `baseDAO.tz` with `trivialDAO` storage is around `18kB`
- `baseDAO.tz` with `registryDAO` storage is around `26kB`
- `baseDAO.tz` with `treasuryDAO` storage is around `23kB`

The previous limit is `16kB`. With the new protocol, the limit is now increased to around `32kB`.

Solution: Remove `morley-large-orginator` from `Makefile`, remove
`originating.md`, replace the mention of it with `tezos-client` in
README, update the tests.