[#256] Allow more xtz entrypoints
Problem: We currently have a safety check for most entrypoints to
disallow xtz tokens to be mistakenly sent to the contract and allow
these funds to be stuck there.
We need to reconsider on which entrypoints we should keep this check
and on which we don't need to (see #91).

Solution: Move `accept_ownership` and `transfer_ownership` to
`allowing_xtz` entrypoints.