[#237] Update contract and docs to use blocks instead of timestamp
Problem: Right now we use timestamps to denote time durations, like
voting periods, and proposal flush and expiry times. But we have to
change it to use the number of blocks.

Solution: Add a single field record to denote durations in number of
blocks and use it in places that currently use timestamps. Also updates
Makefile, specification/build documents.