[#233] Use .tz file as haskell deps
Problem: currently we do not state in the Haskell's build system that
there are `.tz` files we want to depend on, we only call
`qAddDependentFile` (or a function that uses it).
This way the file is indeed fetched, however once the project is built,
`cabal`/`stack` do not know that the module that fetches the contract
or the test storages has to be rebuild when `.tz` files changes.

Solution: add the .tz files to `extra-source-files` section of
`package.yaml` and move `README.md` to a separate `extra-doc-files`
section.
Additionally, remove env var and refactor some code as a consequence.