[#263] Remove lambdas from config and hard code calls to them
Problem: We have to remove the lambdas for proposal_check,
proposal_rejection_slash_value, decision_lambda and custom entrypoints
from config.

Solution: Remove the lambdas from config part of storage. To maintain
the current behavior, we will build separate contract for each of the
variants. This will be accomplished by including the
'implementation.mligo' module, that will be derived from the
corresponding source file for the variant.

For proposal_check, rejected_slash_value and decision_lambda calls, the
three variants share the same function type. So we just need to include
code for the variants using the build time switching mechanism.

For custom entrypoints, the type of the parameter is different
among the variants (all custom entrypoint are added under a single
constructor) and the parameter type part that represent this
entrypoint is also different. So the dynamically invoked
implementation file will also export the part of the parameter that is
variable.

The existing code that call the lambdas and custom entrypoint is
modified to call the function names that are expected to be exported
from the implementation file.