[#314] Improve `storage.start_level`
Problem: The `storage.start_level` is set at the start of origination
instead of when the origination is completed. Because of this, in some
case when the origination takes longer than expected, the
`storage.start_level` is out of sync.

Solution: Provide a way to set `storage.start_level` after the
origination is completed (by calling `Transfer_ownership`). This won't
take effect if the `start_level` is already set previously.
Update the tests, and remove origination workaround.