[#250] fix documentation around rejected proposal slash value
Problem: the documentation wasn't fully updated to reflect the
behavior of the rejected proposal slash value lambda, the last
time it was changed.
As a result the docs contradict each other in some cases and
don't match the actual behavior.

Solution: fix the various issues in the docs related to this
lambda and its usage.