[#263] Amend Makefile and ligo src to support concrete storage variants
Problem: We want to remove polymorphism on contract extra for storage
and wants to generate different concrete storages for each variants.

Solution: Reorganize ligo code to enable definition and inclusion of
implementation specific contract storage. As a part of this, the make
file will also add an 'implementation_storage.mligo' file depending on
the variant that needs building.