[#324] Update Haskell storage type to include config in storage.
Problem: We have remove full_storage type from LIGO code and is
including the config in 'storage' type itself. We have to update the
Haskell code accordingly.

Solution: Include the config type in storage and remove FullStorage
type. As part of this change, we have switch back to using custom
generic derivation. But since

https://gitlab.com/morley-framework/morley/-/issues/754

is still unresolved, we manually derive the rpc counterpart of the
storage.