[#329] Split test module
Problem: Compilation have become very slow due to the additional
constraints on the registry/treasury constraints.

Solution: Until a better solution is found, split the big test module
into separate modules with one test each. This was recompilations during
development shouldn't slow down too much, because only the changed
module and the constriants for a single test within it need to be
recompiled.