[#329] Fix timing of Lambda dao tests
Problem: When we added LambdaDAO, we added an offet to the current level
while setting initial level, and then wait to make sure that the
origination function returns after that level is past. But this waiting
code was missing in the origination function for plain lambdaDAO tests.

Solution: Add the waiting code to the origination funcion.