[Chore] Run nettests in a single thread
Problem: currently, nettests are run with tasty in parallel mode. This
is rather pointless, however, as we currently can only run one nettest
at a time. Besides, even if we could run multiple tests in parallel,
that would make timings less reliable and we already struggle with that.

Solution: explicitly pass `--num-threads=1` to the nettest command.