[Chore] Fix multiple updates in registry's config proposals
Problem: in registryDAO's config proposals multiple updates
to the config can be specified, but due to a bug not all of
them will become effective when the proposal is flushed.

Solution: fix the updates in the code to make config proposals
able again to update multiple config parameters at once.